ACE-532: Moved authentication for servlets to the AceServletContextHelper
This also fixes ACE-256
