+++ org.apache.ace.deployment/src/org/apache/ace/deployment/servlet/AgentDeploymentServlet.java
@@ -18,8 +18,6 @@
  */
 package org.apache.ace.deployment.servlet;
 
-import static javax.servlet.http.HttpServletResponse.SC_UNAUTHORIZED;
-
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
@@ -42,24 +40,21 @@ import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
 
-import org.apache.ace.authentication.api.AuthenticationService;
 import org.apache.ace.connectionfactory.ConnectionFactory;
-import org.apache.felix.dm.Component;
-import org.apache.felix.dm.DependencyManager;
 import org.osgi.framework.Version;
 import org.osgi.service.cm.ConfigurationException;
 import org.osgi.service.cm.ManagedService;
 import org.osgi.service.log.LogService;
-import org.osgi.service.useradmin.User;
 import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
 
 public class AgentDeploymentServlet extends HttpServlet implements ManagedService {
+    private static final long serialVersionUID = 1L;
+
     private static final int BUFFER_SIZE = 1024 * 32;
-    /** A boolean denoting whether or not authentication is enabled. */
-    private static final String KEY_USE_AUTHENTICATION = "authentication.enabled";
+
     /** URL to the OBR that is used for finding versions of the agent. */
     private static final String KEY_OBR_URL = "obr.url";
 
@@ -70,12 +65,9 @@ public class AgentDeploymentServlet extends HttpServlet implements ManagedServic
     public static final String TEXT_MIMETYPE = "text/plain";
 
     // injected by Dependency Manager
-    private volatile DependencyManager m_dm;
     private volatile LogService m_log;
-    private volatile AuthenticationService m_authService;
     private volatile ConnectionFactory m_connectionFactory;
     // See updated()
-    private boolean m_useAuth = false;
     private URL m_obrURL;
 
     private final String m_repositoryXML = "index.xml";
@@ -103,14 +95,6 @@ public class AgentDeploymentServlet extends HttpServlet implements ManagedServic
     @Override
     public void updated(Dictionary<String, ?> settings) throws ConfigurationException {
         if (settings != null) {
-            String useAuthString = (String) settings.get(KEY_USE_AUTHENTICATION);
-            if (useAuthString == null
-                || !("true".equalsIgnoreCase(useAuthString) || "false".equalsIgnoreCase(useAuthString))) {
-                throw new ConfigurationException(KEY_USE_AUTHENTICATION, "Missing or invalid value!");
-            }
-            boolean useAuth = Boolean.parseBoolean(useAuthString);
-            m_useAuth = useAuth;
-
             String obrURL = (String) settings.get(KEY_OBR_URL);
             try {
                 URL url = new URL(obrURL);
@@ -125,20 +109,8 @@ public class AgentDeploymentServlet extends HttpServlet implements ManagedServic
             }
         }
         else {
-            m_useAuth = false;
             m_obrURL = null;
         }
-    }
-
-    /**
-     * Called by Dependency Manager upon initialization of this component.
-     * 
-     * @param comp the component to initialize, cannot be <code>null</code>.
-     */
-    protected void init(Component comp) {
-        comp.add(m_dm.createServiceDependency()
-            .setService(AuthenticationService.class).setRequired(m_useAuth)
-        );
     }
 
     @Override
@@ -165,37 +137,7 @@ public class AgentDeploymentServlet extends HttpServlet implements ManagedServic
 
     protected URLConnection openConnection(URL url) throws IOException {
         return m_connectionFactory.createConnection(url);
-    }
-
-    @Override
-    protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        if (!authenticate(req)) {
-            // Authentication failed; don't proceed with the original request...
-            resp.sendError(SC_UNAUTHORIZED);
-        }
-        else {
-            // Authentication successful, proceed with original request...
-            super.service(req, resp);
-        }
-    }
-
-    /**
-     * Authenticates, if needed the user with the information from the given request.
-     * 
-     * @param request
-     *            the request to obtain the credentials from, cannot be <code>null</code>.
-     * @return <code>true</code> if the authentication was successful, <code>false</code> otherwise.
-     */
-    private boolean authenticate(HttpServletRequest request) {
-        if (m_useAuth) {
-            User user = m_authService.authenticate(request);
-            if (user == null) {
-                m_log.log(LogService.LOG_INFO, "Authentication failure!");
-            }
-            return (user != null);
-        }
-        return true;
-    }
+    }    
 
     private void closeSilently(Closeable resource) {
         try {
