+++ org.apache.ace.deployment/src/org/apache/ace/deployment/servlet/DeploymentServlet.java
@@ -18,12 +18,9 @@
  */
 package org.apache.ace.deployment.servlet;
 
-import static javax.servlet.http.HttpServletResponse.SC_UNAUTHORIZED;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.Dictionary;
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -34,29 +31,21 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.apache.ace.authentication.api.AuthenticationService;
 import org.apache.ace.deployment.processor.DeploymentProcessor;
 import org.apache.ace.deployment.provider.ArtifactData;
 import org.apache.ace.deployment.provider.DeploymentProvider;
 import org.apache.ace.deployment.provider.OverloadedException;
 import org.apache.ace.deployment.streamgenerator.StreamGenerator;
-import org.apache.felix.dm.Component;
-import org.apache.felix.dm.DependencyManager;
 import org.osgi.framework.ServiceReference;
-import org.osgi.service.cm.ConfigurationException;
-import org.osgi.service.cm.ManagedService;
 import org.osgi.service.log.LogService;
-import org.osgi.service.useradmin.User;
 
 /**
  * The DeploymentServlet class provides in a list of versions available for a target and a stream of data containing the
  * DeploymentPackage (or fix package) for a specific target and version.
  */
-public class DeploymentServlet extends HttpServlet implements ManagedService {
+public class DeploymentServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
 
-    /** A boolean denoting whether or not authentication is enabled. */
-    private static final String KEY_USE_AUTHENTICATION = "authentication.enabled";
     /** HTTP header name used for Deployment Package size estimate, in bytes. */
     private static final String HEADER_DPSIZE = "X-ACE-DPSize";
     /** Multiplication factor for the DP size to account for slight changes in file change due to resource processors. */
@@ -71,13 +60,9 @@ public class DeploymentServlet extends HttpServlet implements ManagedService {
     private final ConcurrentMap<String, DeploymentProcessor> m_processors = new ConcurrentHashMap<>();
 
     // injected by Dependency Manager
-    private volatile DependencyManager m_dm;
     private volatile LogService m_log;
     private volatile StreamGenerator m_streamGenerator;
     private volatile DeploymentProvider m_provider;
-    private volatile AuthenticationService m_authService;
-
-    private volatile boolean m_useAuth = false;
 
     public void addProcessor(ServiceReference<DeploymentProcessor> ref, DeploymentProcessor processor) {
         String key = (String) ref.getProperty(PROCESSOR);
@@ -99,23 +84,6 @@ public class DeploymentServlet extends HttpServlet implements ManagedService {
         // we do not log this here again, we already did so in 'addProcessor'
         if (key != null) {
             m_processors.remove(key);
-        }
-    }
-
-    @Override
-    public void updated(Dictionary<String, ?> settings) throws ConfigurationException {
-        if (settings != null) {
-            String useAuthString = (String) settings.get(KEY_USE_AUTHENTICATION);
-            if (useAuthString == null
-                || !("true".equalsIgnoreCase(useAuthString) || "false".equalsIgnoreCase(useAuthString))) {
-                throw new ConfigurationException(KEY_USE_AUTHENTICATION, "Missing or invalid value!");
-            }
-            boolean useAuth = Boolean.parseBoolean(useAuthString);
-
-            m_useAuth = useAuth;
-        }
-        else {
-            m_useAuth = false;
         }
     }
 
@@ -187,49 +155,6 @@ public class DeploymentServlet extends HttpServlet implements ManagedService {
         catch (OverloadedException oe) {
             throw new ServletException(oe);
         }
-    }
-
-    /**
-     * Called by Dependency Manager upon initialization of this component.
-     * 
-     * @param comp
-     *            the component to initialize, cannot be <code>null</code>.
-     */
-    protected void init(Component comp) {
-        comp.add(m_dm.createServiceDependency()
-            .setService(AuthenticationService.class)
-            .setRequired(m_useAuth)
-            );
-    }
-
-    @Override
-    protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        if (!authenticate(req)) {
-            // Authentication failed; don't proceed with the original request...
-            resp.sendError(SC_UNAUTHORIZED);
-        }
-        else {
-            // Authentication successful, proceed with original request...
-            super.service(req, resp);
-        }
-    }
-
-    /**
-     * Authenticates, if needed the user with the information from the given request.
-     * 
-     * @param request
-     *            the request to obtain the credentials from, cannot be <code>null</code>.
-     * @return <code>true</code> if the authentication was successful, <code>false</code> otherwise.
-     */
-    private boolean authenticate(HttpServletRequest request) {
-        if (m_useAuth) {
-            User user = m_authService.authenticate(request);
-            if (user == null) {
-                m_log.log(LogService.LOG_INFO, "Authentication failure!");
-            }
-            return (user != null);
-        }
-        return true;
     }
 
     private long estimateDeploymentPackageSize(HttpServletRequest request, String targetID, String version) throws IOException, OverloadedException, AceRestException {
