+++ org.apache.ace.http/bnd.bnd
@@ -6,5 +6,6 @@
 	osgi.cmpn,\
 	org.apache.felix.dependencymanager,\
 	org.apache.felix.http.api,\
-	org.apache.felix.http.servlet-api
+	org.apache.felix.http.servlet-api,\
+	org.apache.ace.authentication.api;version=latest
 -sub: *.bnd
