+++ org.apache.ace.log/src/org/apache/ace/log/server/servlet/Activator.java
@@ -28,7 +28,6 @@ import java.util.Properties;
 
 import javax.servlet.Servlet;
 
-import org.apache.ace.authentication.api.AuthenticationService;
 import org.apache.ace.log.server.store.LogStore;
 import org.apache.felix.dm.Component;
 import org.apache.felix.dm.DependencyActivatorBase;
@@ -42,9 +41,6 @@ import org.osgi.service.log.LogService;
 public class Activator extends DependencyActivatorBase implements ManagedServiceFactory {
 
     private static final String KEY_LOG_NAME = "name";
-
-    /** A boolean denoting whether or not authentication is enabled. */
-    private static final String KEY_USE_AUTHENTICATION = "authentication.enabled";
 
     private final Map<String, Component> m_instances = new HashMap<>(); // String -> Service
     private DependencyManager m_manager;
@@ -81,12 +77,6 @@ public class Activator extends DependencyActivatorBase implements ManagedService
             throw new ConfigurationException(KEY_LOG_NAME, "Log name has to be specified: " + name);
         }
         
-        String useAuthString = (String) dict.get(KEY_USE_AUTHENTICATION);
-        if (useAuthString == null
-            || !("true".equalsIgnoreCase(useAuthString) || "false".equalsIgnoreCase(useAuthString))) {
-            throw new ConfigurationException(KEY_USE_AUTHENTICATION, "Missing or invalid value: " + useAuthString);
-        }
-        boolean useAuth = Boolean.parseBoolean(useAuthString);
         //TODO: Is this the best way ? 
         ((Dictionary)dict).put(HTTP_WHITEBOARD_CONTEXT_SELECT, ACE_WHITEBOARD_CONTEXT_SELECT_FILTER);
 
@@ -94,8 +84,7 @@ public class Activator extends DependencyActivatorBase implements ManagedService
         if (service == null) {
             service = m_manager.createComponent()
                 .setInterface(Servlet.class.getName(), dict)
-                .setImplementation(new LogServlet(name, useAuth))
-                .add(createServiceDependency().setService(AuthenticationService.class).setRequired(useAuth))
+                .setImplementation(new LogServlet(name))
                 .add(createServiceDependency().setService(LogService.class).setRequired(false))
                 .add(createServiceDependency().setService(LogStore.class, "(&("+Constants.OBJECTCLASS+"="+LogStore.class.getName()+")(name=" + name + "))").setRequired(true));
