+++ org.apache.ace.log/src/org/apache/ace/log/server/servlet/LogServlet.java
@@ -18,29 +18,24 @@
  */
 package org.apache.ace.log.server.servlet;
 
-import static javax.servlet.http.HttpServletResponse.SC_UNAUTHORIZED;
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.servlet.ServletException;
 import javax.servlet.ServletInputStream;
 import javax.servlet.ServletOutputStream;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.apache.ace.authentication.api.AuthenticationService;
 import org.apache.ace.feedback.Descriptor;
 import org.apache.ace.feedback.Event;
 import org.apache.ace.feedback.LowestID;
 import org.apache.ace.log.server.store.LogStore;
 import org.apache.ace.range.SortedRangeSet;
 import org.osgi.service.log.LogService;
-import org.osgi.service.useradmin.User;
 
 /**
  * This class acts as a servlet and handles the log protocol. This means a number of requests will be handled:
@@ -88,14 +83,11 @@ public class LogServlet extends HttpServlet {
     // injected by Dependency Manager
     private volatile LogService m_log;
     private volatile LogStore m_store;
-    private volatile AuthenticationService m_authService;
 
     private final String m_name;
-    private final boolean m_useAuth;
 
-    public LogServlet(String name, boolean useAuth) {
+    public LogServlet(String name) {
         m_name = name;
-        m_useAuth = useAuth;
     }
 
     @Override
@@ -153,34 +145,6 @@ public class LogServlet extends HttpServlet {
                 m_log.log(LogService.LOG_WARNING, "Exception trying to close stream after request: " + request.getRequestURL(), ex);
             }
         }
-    }
-
-    @Override
-    protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        if (!authenticate(req)) {
-            // Authentication failed; don't proceed with the original request...
-            resp.sendError(SC_UNAUTHORIZED);
-        } else {
-            // Authentication successful, proceed with original request...
-            super.service(req, resp);
-        }
-    }
-
-    /**
-     * Authenticates, if needed the user with the information from the given request.
-     * 
-     * @param request the request to obtain the credentials from, cannot be <code>null</code>.
-     * @return <code>true</code> if the authentication was successful, <code>false</code> otherwise.
-     */
-    private boolean authenticate(HttpServletRequest request) {
-        if (m_useAuth) {
-            User user = m_authService.authenticate(request);
-            if (user == null) {
-                m_log.log(LogService.LOG_INFO, "Authentication failure!");
-            }
-            return (user != null);
-        }
-        return true;
     }
 
     // Handle a call to the query 'command'
