+++ org.apache.ace.log/test/org/apache/ace/log/server/servlet/LogServletTest.java
@@ -53,7 +53,7 @@ public class LogServletTest {
 
     @BeforeMethod(alwaysRun = true)
     protected void setUp() throws Exception {
-        m_logServlet = new LogServlet("test", false /* useAuth */);
+        m_logServlet = new LogServlet("test");
         TestUtils.configureObject(m_logServlet, LogService.class);
         m_mockStore = new MockLogStore();
         TestUtils.configureObject(m_logServlet, LogStore.class, m_mockStore);
