+++ run-server-allinone/conf/org.apache.ace.deployment.servlet.agent.cfg
@@ -2,4 +2,3 @@
 
 # OBR settings
 obr.url = http://${org.apache.ace.obr}/obr/
-authentication.enabled = false
