--- run-server-allinone/conf/org.apache.ace.gogo.servlet.cfg
@@ -1,3 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under the terms of ASLv2 (http://www.apache.org/licenses/LICENSE-2.0).
-
-authentication.enabled=false
